requirements.txt
Django>=5.1,<5.2
djangorestframework>=3.15,<3.16
django-environ>=0.11,<0.12
psycopg[binary]>=3.2,<3.3
celery>=5.4,<5.5
redis>=5.0,<5.1
django-redis>=5.4,<5.5
channels>=4.1,<4.2
channels-redis>=4.2,<4.3
daphne>=4.1,<4.2
django-ratelimit>=4.1,<4.2
django-cors-headers>=4.4,<4.5
Pillow>=10.4,<10.5
